Added more support to logging
